+++ plugins/Journal/src/Events/JournalEntryCreated.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace LifeWheel\Plugins\Journal\Events;
+
+use App\Models\User;
+
+final readonly class JournalEntryCreated
+{
+    public function __construct(
+        public User $user,
+        public int $entryId,
+    ) {
+    }
+}